Update README.md
The introduction of our medicine ordering website marks a pivotal moment in the realm of healthcare accessibility. In an era dominated by digital solutions, the need for a streamlined method of procuring medications has become increasingly apparent. Our website seeks to address this need by providing a user-friendly platform that revolutionizes the way individuals access essential medications. By leveraging technology, we aim to bridge the gap between patients and pharmacies, offering a convenient and efficient alternative to traditional brick-and-mortar establishments.